feat : 페이지번호가 범위를 벗어나면 에러페이지를 보여주도록 구현하라
- 해당 기능 구현에 앞서, 이전버튼과 다음버튼까지 테스트할 수 있도록 수정했다.
- 추가로 페이지네이션에 관한 변수타입을 long에서 int로 수정했다
- 페이지 범위에 벗어나는 경우에 해당하는 예외를 만들고, BAD_REQUEST로서 처리될 수 있도록 설정했다

+++ src/main/java/com/codessquad/qna/web/domain/QuestionRepository.java
@@ -12,5 +12,5 @@ public interface QuestionRepository extends PagingAndSortingRepository<Question,
 
     Optional<Question> findByIdAndDeletedFalse(Long id);
 
-    Long countAllByDeletedFalse();
+    int countAllByDeletedFalse();
 }
